Change card & button text; insert correct links
The buttons in the three cards at the bottom of the page now point to the
correct URLs. I intentionally made the leftmost card description
text longer than it needs to be (i.e., "How do I get started with
Linux?" could just as easily have been "How do I use Linux?"), since the
cards are evenly spaced apart from one another and their width is
determined by the text length. In the future, it would be better if the
space between the cards (and maybe their width) was manually set.